bug: Bash users get stuck on lesson 7 step 15
Reason: _git_branch_status processed the output of `git status` with a
pipeline, setting $REPLY as a side-effect.  In Bash the RHS of a
pipeline *always* executes in a subshell, which loses the value in
$REPLY.  Zsh is able to do this processing in the environment of the
original shell, which is why this bug went unnoticed for so long.

Solution: process the output with a for loop instead.  Set IFS to \n
to preserve linebreaks in the output of `git status`.